Fix LotSizeFilter json unmarshalling

+++ rest/result.go
@@ -168,9 +168,9 @@ type PriceFilter struct {
 }
 
 type LotSizeFilter struct {
-	MaxTradingQty int `json:"max_trading_qty"`
-	MinTradingQty int `json:"min_trading_qty"`
-	QtyStep       int `json:"qty_step"`
+	MaxTradingQty float64 `json:"max_trading_qty"`
+	MinTradingQty float64 `json:"min_trading_qty"`
+	QtyStep       float64 `json:"qty_step"`
 }
 
 type SymbolInfo struct {
